Changing methods responsible for showing commands
Methods inside CommandWords and Parser classes are right now returning String with valid commands instead of printing valid commands. It gives us opportunity to easier improve our Game later - if we would like to create a GUI instead of printing everything in command line, we would only have to change one class, which is responsible for interaction with user.